+++ modelica-export/src/main/java/eu/itesla_project/modelica_export/records/DanglingLineRecord.java
@@ -50,7 +50,7 @@ public class DanglingLineRecord extends ModelicaRecord {
         Equipments.ConnectionInfo info1 = Equipments.getConnectionInfoInBusBreakerView(this.danglingLine.getTerminal());
         Bus b1 = info1.getConnectionBus();
 
-        if (!Float.isNaN(b1.getV()) && info1.isConnected()) {
+        if (!Double.isNaN(b1.getV()) && info1.isConnected()) {
             if (super.isCorrect()) {
                 if (super.getModelicaType() != null) {
                     this.addValue(super.getModelicaType() + StaticData.WHITE_SPACE);
@@ -131,14 +131,14 @@ public class DanglingLineRecord extends ModelicaRecord {
      */
     private void setParameters(float snref) {
         //this.iidmbranchParameters = new ArrayList<IIDMParameter>();
-        float tNominalV = this.danglingLine.getTerminal().getVoltageLevel().getNominalV();
-        float voltage = !Float.isNaN(tNominalV) ? tNominalV : 0;
-        float z = (voltage * voltage) / snref;
+        double tNominalV = this.danglingLine.getTerminal().getVoltageLevel().getNominalV();
+        double voltage = !Double.isNaN(tNominalV) ? tNominalV : 0;
+        float z = (float) (voltage * voltage) / snref;
 
-        super.addParameter(this.iidmbranchParameters, StaticData.R, this.danglingLine.getR() / z);
-        super.addParameter(this.iidmbranchParameters, StaticData.X, this.danglingLine.getX() / z);
-        super.addParameter(this.iidmbranchParameters, StaticData.G, this.danglingLine.getG() * z);
-        super.addParameter(this.iidmbranchParameters, StaticData.B, this.danglingLine.getB() * z);
+        super.addParameter(this.iidmbranchParameters, StaticData.R, (float) this.danglingLine.getR() / z);
+        super.addParameter(this.iidmbranchParameters, StaticData.X, (float) this.danglingLine.getX() / z);
+        super.addParameter(this.iidmbranchParameters, StaticData.G, (float) this.danglingLine.getG() * z);
+        super.addParameter(this.iidmbranchParameters, StaticData.B, (float) this.danglingLine.getB() * z);
     }
 
     @Override
